Floating Stoic and IPR/IRR Units
Improved capability for stoichiometry floating.  Stoic should be a product if the species only appears as a product and a reactant if the species only appears as a reactant.  This functionality was failing for hourly sums, but has now been fixed.

IPR/IRR variables can now carry properties as attributes of the array

git-svn-id: https://dawes.sph.unc.edu/uncaqmlsvn/net_balance/trunk@178 f2e443aa-f90a-0410-907f-bad4a36f275b