Improving mechanism import ease
Mechanisms were previously imported using a module with awareness of available mechanisms and internal directory structure.  A typical work flow would be

> from net_balance.getmech import get_pure_mech
> mech = get_pure_mech('cb05_cmaq')
> ...

The new import method requires each mechanism to have its own initialization call in its __init__.py file.  A typical work flow would be

> from net_balance.mechanisms.cb05_cmaq import mech
> ...

The improvement relates to previous requirement that the user be familiar with an arbitrarily named function.  The new method allows users to dynamically query the mechanisms module for available mechanisms with the caveat that all mechanisms implement the mech call.

git-svn-id: https://dawes.sph.unc.edu/uncaqmlsvn/net_balance/trunk@181 f2e443aa-f90a-0410-907f-bad4a36f275b