Process Graphing
Addied process graphing and attempted to make it incrementally useful.  The default graph includes all individual processes with a legend on the right.  The processes can be selected by adding a keyword argument that corresponds to the process.  The keyword argument is a dictionary that includes line properties or is empty.

Added process evalution error checking

Added optional IPR and IRR functionality.

Improved Species Definitions

Species are now defined using a NAME = COMPONENT1 + COMPONENT2... pattern.

git-svn-id: https://dawes.sph.unc.edu/uncaqmlsvn/net_balance/trunk@215 f2e443aa-f90a-0410-907f-bad4a36f275b