Refactoring Species Addition
Addition of species was taking too long.  The problem was incremental addition of long species families.  Instead of a long chain of additions S1+S2+S3+S...+SN, I have added a species sum function.  A sum function reduces the overhead in individual addition calls.  This is a big save\!

git-svn-id: https://dawes.sph.unc.edu/uncaqmlsvn/net_balance/trunk@218 f2e443aa-f90a-0410-907f-bad4a36f275b